Move bfp kernels into correct folder
