+++ aie_kernels/aie2p/mm_bfp_mixed.cc
@@ -8,7 +8,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "../../aie_kernels/aie_kernel_utils.h"
+#include "../aie_kernel_utils.h"
 #include <aie_api/aie.hpp>
 
 template <typename T, int M, int N>
